lambda-functions/create_user/requirements.txt
boto3==1.35.71
DateTime==5.5
python-gitlab==5.1.0
requests==2.32.3
